+++ src/models/__init__.py
@@ -3,5 +3,9 @@ WRGD data models.
 """
 
 from .coordinate import Coordinate
+from .road_statistics import RoadStatistics
 
-__all__ = ["Coordinate"]
+__all__ = [
+    "Coordinate",
+    "RoadStatistics",
+]